move ie tests to nightly set
